pyproject.toml
[tool.poetry]
name = "rps-backend"
version = "0.1.0"
description = ""
authors = ["archdrdr <archdroider@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
email-validator = "^2.2.0"
python-multipart = "^0.0.9"
fastapi-cli = "^0.0.5"
pydantic-settings = "^2.4.0"
pydantic-extra-types = "^2.9.0"
sqlalchemy = "^2.0.32"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
ruff = "^0.6.2"
types-passlib = "^1.7.7.20240819"


[tool.poetry.group.test.dependencies]
httpx = "^0.27.1"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Pytest settings
[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"


# Mypy settings
[tool.mypy]
python_version = 3.12
disallow_untyped_defs = true
packages = ["rps_backend", "tests"]

# Ruff settings
[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

[tool.ruff.format]
quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"
